requirements.txt
Django==4.2.4
django-cors-headers==4.2.0
djangorestframework==3.14.0
djoser==2.2.0
dj-database-url==2.0.0 
python-dotenv==1.0.0
pytz==2023.3
whitenoise==6.5.0
psycopg2==2.9.7
gunicorn==21.2.0
django-storages==1.13.2